refactor: move create deployment to specific file
