+++ api/api/versions.json
@@ -3,7 +3,7 @@
     {
       "version": "v1",
       "status": "active",
-      "release_date": "2025-08-19T22:07:00.891959687+05:30",
+      "release_date": "2025-08-19T22:09:27.236125234+05:30",
       "end_of_life": "0001-01-01T00:00:00Z",
       "changes": [
         "Initial API version"
